requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
Werkzeug==2.3.7
openai==1.3.0
pydub==0.25.1
PyYAML==6.0.1
faiss-cpu==1.7.4
tiktoken==0.5.2
numpy==1.24.3
supabase==2.0.0
